+++ server/service/conditions/Condition.php
@@ -21,6 +21,11 @@ class Condition
     private $user_input;
 
     /**
+     * The router instance which is used to generate valid links.
+     */
+    private $router;
+
+    /**
      * Start the session.
      *
      * @param object $db
@@ -28,10 +33,11 @@ class Condition
      * @param object $user_input
      *  The user_input instance which grants access to the user_input.
      */
-    public function __construct($db, $user_input)
+    public function __construct($db, $user_input, $router)
     {
         $this->db = $db;
         $this->user_input = $user_input;
+        $this->router = $router;
     }
 
     /**
@@ -86,6 +92,8 @@ class Condition
         $j_condition = str_replace('__current_date__', date('Y-m-d'), $j_condition); // replace __current_date__
         $j_condition = str_replace('__current_date_time__', date('Y-m-d H:i'), $j_condition); // replace __current_date_time__
         $j_condition = str_replace('__current_time__', date('H:i'), $j_condition); // replace __current_time__
+        $keyword = $this->router->get_keyword_from_url();
+        $j_condition = str_replace('__keyword__', $keyword, $j_condition); // replace __keyword__
         // replace form field keywords with the actual values.
         $pattern = '~"' . $this->user_input->get_input_value_pattern() . '"~';
         preg_match_all($pattern, $j_condition, $matches, PREG_PATTERN_ORDER);